Changed first subsection of terms.md to include a contact email.
In the first subsection about accepting the privacy, I included an email address instead of just "please contact us." I doubt most, if any, users will have problems with the terms. But if they do, it should be easy to find the correct contact address through which to reach out and receive prompt response. Maybe best to list it here?